requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==1.10.13

supabase==2.4.0
gotrue==2.5.4
postgrest==0.16.8
storage3==0.7.7
realtime==1.0.6
httpx==0.27.0
